extract ScoreBound and SortedSetOrder into sorted_set_common module
