+++ src/cache/messages/data/sorted_set/sorted_set_fetch_by_rank.rs
@@ -4,20 +4,9 @@ use momento_protos::common::Unbounded;
 
 use crate::cache::messages::data::sorted_set::sorted_set_fetch_response::SortedSetFetchResponse;
 use crate::cache::messages::MomentoRequest;
+use crate::cache::SortedSetOrder;
 use crate::utils::prep_request_with_timeout;
 use crate::{CacheClient, IntoBytes, MomentoResult};
-
-/// The order with which to sort the elements by score in the sorted set.
-/// The sort order determines the rank of the elements.
-/// The elements with same score are ordered lexicographically.
-#[repr(i32)]
-#[derive(Debug, Copy, Clone, PartialEq)]
-pub enum SortedSetOrder {
-    /// Scores are ordered from low to high. This is the default order.
-    Ascending = 0,
-    /// Scores are ordered from high to low.
-    Descending = 1,
-}
 
 /// Request to fetch the elements in a sorted set by their rank.
 ///
